Use str#find in NamedParameter's match statement
(in impl Node<T> find)
For the match value in NodeKind::Parameter we need to determine the byte
index, **not** the char index, otherwise we will be off for multibyte
characters in the slice further down. In worst case scenarios this could
lead to panics due to the index not being on a char boundary for the slice.

Example:

```
// .chars().position(...) vs .find(...)

fn main() {
    let p1 = "sx/";
    let p2 = "ßx/";
    let p3 = "xß/";

    check(p1);
    check(p2);
    check(p3);
}

fn check(p: &str) {
    println!(
      "p: {:?} | {:?} | byte len: {}, char count: {}",
      p, p.as_bytes(), p.len(), p.chars().count());

    let i = p.find('/').unwrap();
    println!("find / => {:?}", i);
    println!("p[i..] => {:?}", &p[i..]);

    let i = p.chars().position(|x| x == '/').unwrap();
    println!("c.p? / => {:?}", i);
    if p.is_char_boundary(i) {
        println!("p[i..] => {:?}", &p[i..]);
    } else {
        println!("p[i..] => would panic (not at char boundary)!");
    }
}
```

References:
- <https://doc.rust-lang.org/std/primitive.str.html#method.find>